+++ src/moldybreadpkg/fedora.nim
@@ -227,6 +227,15 @@ method download(this: FedoraRecord, output_directory: string, suffix=""): bool {
   else:
     false
 
+method download_page_with_relationship(this: FedoraRecord, output_directory, book_pid, page_number: string): bool {. base .} =
+  let response = this.client.request(this.uri, httpMethod = HttpGet)
+  if response.status == "200 OK":
+    let extension = this.get_extension(response.headers)
+    discard this.write_output(fmt"{book_pid}_{page_number}{extension}", response.body, output_directory)
+    true
+  else:
+    false
+
 method get_content_model(this: FedoraRecord): string {. base .} =
   let response = this.client.request(this.uri, httpMethod = HttpGet)
   if response.status == "200 OK":
@@ -414,13 +423,23 @@ method determine_pages(this: FedoraRequest): seq[string] {. base .} =
       bar.increment()
   bar.finish()
 
-method find_book_and_page(this: FedoraRequest): seq[(string, string, string)] {. base .} =
+method download_page_with_book_relationship*(this: FedoraRequest, datastream_id: string): seq[(string, string, string)] {. base .} =
+  ## Downloads parts of pages of a book and names it based on the relationship.
+  ##
+  ## Example:
+  ##
+  ## .. code-block:: nim
+  ##
+  ##    let fedora_connection = initFedoraRequest(pid_part="test", output_directory="output")
+  ##    fedora_connection.results = fedora_connection.populate_results()
+  ##    echo fedora_connection.download_page_with_book_relationship("OBJ")
+  ##
   let predicate = "&predicate=info%3afedora%2ffedora-system%3adef%2frelations-external%23isMemberOf"
   var
     pid: string
     bar = newProgressBar(total=len(this.results), step=int(ceil(len(this.results)/100)))
   let ticks = progress_prep(len(this.results))
-  echo "\n\nGetting Matching Books with Page and PID:\n"
+  echo "\n\nDownloading Pages and Naming as Book_PageNumber.extension:\n"
   bar.start()
   for i in 1..len(this.results):
     pid = this.results[i-1]
@@ -432,6 +451,7 @@ method find_book_and_page(this: FedoraRequest): seq[(string, string, string)] {.
         book = newTriple(response).obj.replace("<info:fedora/", "").replace(">", "")
         page_triple = FedoraRecord(client: this.client, uri: fmt"{this.base_url}/fedora/objects/{pid}/relationships?subject=info%3afedora%2f{pid}&format=turtle&predicate=http://islandora.ca/ontology/relsext%23isPageNumber", pid: pid).get()
         page = newTriple(page_triple).obj.replace(""""""", "")
+      discard FedoraRecord(client: this.client, uri: fmt"{this.base_url}/fedora/objects/{pid}/datastreams/{datastream_id}/content", pid: pid).download_page_with_relationship(this.output_directory, book, page)
       result.add((book, page, pid))
     if i in ticks:
       bar.increment()
@@ -449,6 +469,7 @@ method harvest_datastream_no_pages*(this: FedoraRequest, datastream_id="MODS"): 
   ##    let fedora_connection = initFedoraRequest(pid_part="test")
   ##    fedora_connection.results = fedora_connection.populate_results()
   ##    discard fedora_connection.harvest_datastream_no_pages("DC")
+  ##
   var
     not_pages = this.determine_pages()
     successes, errors: seq[string]
@@ -1083,6 +1104,6 @@ method find_xacml_restrictions*(this: FedoraRequest): seq[(string, seq[XACMLRule
   bar.finish()
 
 when isMainModule:
-  let fedora_connection = initFedoraRequest(pid_part="test")
+  let fedora_connection = initFedoraRequest(pid_part="test", output_directory="output")
   fedora_connection.results = fedora_connection.populate_results()
-  echo fedora_connection.find_book_and_page()
+  echo fedora_connection.download_page_with_book_relationship("OBJ")